Moving ALL deprecated functions into Deprecated.js. They import the new classes and use them. Can be easily removed later. Also removed CommandGenerator class and just have 3 const functions exported now
