+++ lib/Connection.js
@@ -176,59 +176,4 @@ class Connection {
   }
 }
 
-// Deprecated
-class iConn {
-  constructor(db, user, pwd, option) {
-    this.connection = new Connection(db, user, pwd, option);
-  }
-
-  /**
-   * @description override the default timeout value for sync mode.
-   * @param {number} seconds
-   */
-  setTimeout(seconds) {
-    this.connection.setTimeout(seconds);
-  }
-
-  /**
-   * @description
-   * enables or disables the verbose mode for debugging
-   * returns the current state of debug flag
-   * @param {boolean} [flag]
-   * @returns {boolean} the current state of the debug flag
-   */
-  debug(flag) {
-    return this.connection.debug(flag);
-  }
-
-
-  /**
-   * @description returns conn property from iConn object.
-   * @returns {object} the conn property from iConn object.
-   */
-  getConnection() {
-    return this.connection.getConnection();
-  }
-
-  /**
-   * @description adds XML request to command list
-   * @param {string | object} xml
-   */
-  add(xml) {
-    this.connection.add(xml);
-  }
-
-  /**
-   * @description
-   * Invokes transport with XML input from command list
-   * Calls user provided callback with the XML output
-   * @param {fuction} callback
-   * @param {boolean} sync
-   */
-  run(callback) {
-    this.connection.run(callback);
-  }
-}
-
 module.exports.Connection = Connection;
-module.exports.iConn = iConn;
