+++ lib/ProgramCall.js
@@ -105,52 +105,4 @@ class ProgramCall {
   }
 }
 
-
-// DEPRECATED: For now, routes calls to ProgramCall, but will be removed at a later time.
-class iPgm {
-  /**
-   * @description Creates a new iPgm object
-   * @constructor
-   * @param {string} pgm
-   * @param {object} [options]
-   */
-  constructor(pgm, options) {
-    console.warn('As of v1.0, class \'iPgm\' is deprecated. Please use \'ProgramCall\' instead.');
-    this.ProgramCall = new ProgramCall(pgm, options);
-  }
-
-  /**
-    * @description adds a parameter to the program XML
-    * @param {string | array} data
-    * @param {string} [type]
-    * @param {object} options
-    * @param {*} inDs
-    */
-  addParam(data, type, options, inDs) {
-    console.warn('As of v1.0, \'iPgm.addParam(data, type, options, inDs)\' is deprecated. Please use \'ProgramCall.addParam(data, type, options, inDs)\' instead.');
-    return this.ProgramCall.addParam(data, type, options, inDs);
-  }
-
-  /**
-   * @description adds a return element to the program XML
-   * @param {string} data
-   * @param {string} type
-   * @param {object} [options]
-   */
-  addReturn(data, type, options) {
-    console.warn('As of v1.0, \'iPgm.addReturn(data, type, options)\' is deprecated. Please use \'ProgramCall.addParam(data, type, options)\' instead.');
-    return this.ProgramCall.addReturn(data, type, options);
-  }
-
-  /**
-   * @description returns the current program XML
-   * @returns {string} the generated program XML
-   */
-  toXML() {
-    console.warn('As of v1.0, \'iPgm.toXML()\' is deprecated. Please use \'ProgramCall.toXML()\' instead.');
-    return this.ProgramCall.toXML();
-  }
-}
-
 module.exports.ProgramCall = ProgramCall;
-module.exports.iPgm = iPgm;
