+++ lib/SqlCall.js
@@ -344,205 +344,4 @@ class SqlCall {
   }
 }
 
-
-class iSql {
-/**
-   * @description Creates a new iSql object
-   * @constructor
-   */
-  constructor() {
-    this.sqlCall = new SqlCall();
-  }
-
-  /**
-   * @description adds sql connect XML
-   * @param {object} [options]
-   */
-  connect(options) {
-    this.sqlCall.connect(options);
-  }
-
-  /**
-   * @description adds sql options XML
-   * @param {object} options
-   */
-  setOptions(options) {
-    this.sqlCall.setOptions(options);
-  }
-
-  /**
-   * @description adds sql query XML
-   * @param {string} stmt
-   * @param {object} [options]
-   */
-  addQuery(stmt, options) {
-    this.sqlCall.addQuery(stmt, options);
-  }
-
-  /**
-   * @description adds sql prepare XML
-   * @param {string} stmt
-   * @param {object} [options]
-   */
-  prepare(stmt, options) {
-    this.sqlCall.prepare(stmt, options);
-  }
-
-  /**
-   * @description adds sql execute XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  execute(params, options) {
-    this.sqlCall.execute(params, options);
-  }
-
-  /**
-   * @description adds sql table XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  tables(params, options) {
-    this.sqlCall.tables(params, options);
-  }
-
-  /**
-   * @description adds sql table priv XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  tablePriv(params, options) {
-    this.sqlCall.tablePriv(params, options);
-  }
-
-  /**
-   * @description adds sql columns XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  columns(params, options) {
-    this.sqlCall.columns(params, options);
-  }
-
-  /**
-   * @description adds sql special XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  special(params, options) {
-    this.sqlCall.special(params, options);
-  }
-
-  /**
-   * @description adds sql column priv XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  columnPriv(params, options) {
-    this.sqlCall.columnPriv(params, options);
-  }
-
-  /**
-   * @description adds sql procedures XML
-   * @param {*} params
-   * @param {*} [options]
-   */
-  procedures(params, options) {
-    this.sqlCall.procedures(params, options);
-  }
-
-  /**
-   * @description adds sql pcolumns XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  pColumns(params, options) {
-    this.sqlCall.pColumns(params, options);
-  }
-
-  /**
-   * @description adds sql primary keys XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  primaryKeys(params, options) {
-    this.sqlCall.primaryKeys(params, options);
-  }
-
-  /**
-   * @description adds sql foriegn keys XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  foreignKeys(params, options) {
-    this.sqlCall.foreignKeys(params, options);
-  }
-
-  /**
-   * @description adds sql stats XML
-   * @param {array} params
-   * @param {object} [options]
-   */
-  statistics(params, options) {
-    this.sqlCall.statistics(params, options);
-  }
-
-  /**
-   * @description adds sql commit XML
-   * @param {object} options
-   */
-  commit(options) {
-    this.sqlCall.commit(options);
-  }
-
-  /**
-   * @description adds sql row count XML
-   * @param {object} options
-   */
-  rowCount(options) {
-    this.sqlCall.rowCount(options);
-  }
-
-  /**
-   * @description adds sql row count XML
-   * @param {object} options
-   */
-  count(options) {
-    this.sqlCall.count(options);
-  }
-
-  /**
-   * @description adds sql describe XML
-   * @param {object} options
-   */
-  describe(options) {
-    this.sqlCall.describe(options);
-  }
-
-  /**
-   * @description adds sql fetch XML
-   * @param {object} options
-   */
-  fetch(options) {
-    this.sqlCall.fetch(options);
-  }
-
-  /**
-   * @description adds sql free XML
-   */
-  free() {
-    this.sqlCall.free();
-  }
-
-  /**
-   * @description returns the current sql XML
-   * @returns {string} - the generted sql XML
-   */
-  toXML() {
-    return this.sqlCall.toXML();
-  }
-}
-
-// iSql is the deprecated version of the class. Removed at a later time.
-module.exports.iSql = iSql;
 module.exports.SqlCall = SqlCall;
