+++ lib/Toolkit.js
@@ -1750,173 +1750,4 @@ class Toolkit {
   }
 }
 
-// This file contains the classes that are deprecated in v1.0. The implementations for all of the
-// functions in these classes prior to v1.0 are moved into the Toolkit class. The classes contained
-// in this file are now wrappers for those Toolkit class calls to enable backward compatability with
-// previous versions, but will throw a warning message. When we no longer want to support calling
-// these methods through these deprecated class names, we can simple disable this file.
-
-class iDataQueue {
-  constructor(conn) {
-    console.warn('As of v1.0, class \'iDataQueue\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
-  }
-
-  sendToDataQueue(name, lib, data, cb) {
-    console.warn('As of v1.0, \'iDataQueue\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.sendToDataQueue(name, lib, data, cb);
-  }
-
-  receiveFromDataQueue(name, lib, length, cb) {
-    console.warn('As of v1.0, \'iDataQueue\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.receiveFromDataQueue(name, lib, length, cb);
-  }
-
-  clearDataQueue(name, lib, cb) {
-    console.warn('As of v1.0, \'iDataQueue\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.clearDataQueue(name, lib, cb);
-  }
-}
-
-class iNetwork {
-  constructor(conn) {
-    console.warn('As of v1.0, class \'iNetwork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
-  }
-
-  getTCPIPAttr(cb) {
-    console.warn('As of v1.0, \'iNetwork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getTCPIPAttr(cb);
-  }
-
-  getNetInterfaceData(ipaddr, cb) {
-    console.warn('As of v1.0, \'iNetwork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getNetInterfaceData(ipaddr, cb);
-  }
-}
-
-class iObj {
-  constructor(conn) {
-    console.warn('As of v1.0, class \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
-  }
-
-  retrUsrAuth(usr, type, obj, lib = '*LIBL', cb) {
-    console.warn('As of v1.0, \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.retrUsrAuth(usr, type, obj, lib, cb);
-  }
-
-  retrCmdInfo(cmd, lib = '*LIBL', cb) {
-    console.warn('As of v1.0, \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.retrCmdInfo(cmd, lib, cb);
-  }
-
-  retrPgmInfo(pgm, lib = '*LIBL', cb) {
-    console.warn('As of v1.0, \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.retrPgmInfo(pgm, lib, cb);
-  }
-
-  retrSrvPgmInfo(srvpgm, lib = '*LIB', cb) {
-    console.warn('As of v1.0, \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.retrSrvPgmIfno(srvpgm, lib, cb);
-  }
-
-  retrUserInfo(user, cb) {
-    console.warn('As of v1.0, \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.retrUserIfno(user, cb);
-  }
-
-  retrUserAuthToObj(path, cb) {
-    console.warn('As of v1.0, \'iObj\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.retrUserAuthToObj(path, cb);
-  }
-}
-
-class iProd {
-  constructor(conn) {
-    console.warn('As of v1.0, class \'iProd\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
-  }
-
-  getPTFInfo(PTFID, cb) {
-    console.warn('As of v1.0, \'iProd\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getPTFInfo(PTFID, cb);
-  }
-
-  getProductInfo(prodID, option, cb) {
-    console.warn('As of v1.0, \'iProd\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getProductInfo(prodID, option, cb);
-  }
-
-  getInstalledProducts(cb) {
-    console.warn('As of v1.0, \'iProd\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getInstalledProducts(cb);
-  }
-}
-
-class iUserSpace {
-  constructor(conn) {
-    console.warn('As of v1.0, class \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
-  }
-
-  createUserSpace(name, lib, attr, size, auth, desc, cb) {
-    console.warn('As of v1.0, \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.createUserSpace(name, lib, attr, size, auth, desc, cb);
-  }
-
-  setUserSpaceData(name, lib, length, msg, cb) {
-    console.warn('As of v1.0, \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.setUserSpaceData(name, lib, length, msg, cb);
-  }
-
-  getUserSpaceData(name, lib, length, cb) {
-    console.warn('As of v1.0, \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getUserSpaceData(name, lib, length, cb);
-  }
-
-  deleteUserSpace(name, lib, cb) {
-    console.warn('As of v1.0, \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.deleteUserSpace(name, lib, cb);
-  }
-}
-
-class iWork {
-  constructor(conn) {
-    console.warn('As of v1.0, class \'iWork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
-  }
-
-  getSysValue(sysValue, cb) {
-    console.warn('As of v1.0, \'iWork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getSysValue(sysValue, cb);
-  }
-
-  getSysStatus(cb) {
-    console.warn('As of v1.0, \'iWork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getSysStatus(cb);
-  }
-
-  getSysStatusExt(cb) {
-    console.warn('As of v1.0, \'iWork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getSysStatysExt(cb);
-  }
-
-  getJobStatus(jobId, cb) {
-    console.warn('As of v1.0, \'iWork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getJobStatus(jobId, cb);
-  }
-
-  getDataArea(lib, area, length, cb) {
-    console.warn('As of v1.0, \'iWork\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getDataArea(lib, area, length, cb);
-  }
-}
-
 module.exports.Toolkit = Toolkit;
-module.exports.iDataQueue = iDataQueue;
-module.exports.iNetwork = iNetwork;
-module.exports.iObj = iObj;
-module.exports.iProd = iProd;
-module.exports.iUserSpace = iUserSpace;
-module.exports.iWork = iWork;
