+++ lib/commands.js
@@ -25,59 +25,9 @@ const {
   iXmlNodeShClose,
 } = require('./ixml');
 
-class CommandGenerator {
-  static generateCmdXML(cmd, options) {
-    const xexec = (cmd.indexOf('?') > 0) ? 'rexx' : 'cmd'; // if need return value, use 'rexx'
-    let xml;
-
-    if (options) {
-      xml = iXmlNodeCmdOpen(options.exec, options.hex, options.before, options.after, options.error)
-            + cmd + iXmlNodeCmdClose();
-    } else {
-      xml = iXmlNodeCmdOpen(xexec, '', '', '', '', '') + cmd + iXmlNodeCmdClose();
-    }
-
-    return xml;
-  }
-
-  static generateQshXML(qsh, options) {
-    let xml;
-    if (options) {
-      xml = iXmlNodeQshOpen(options.rows, options.hex, options.before, options.after, options.error)
-            + qsh + iXmlNodeQshClose();
-    } else {
-      xml = iXmlNodeQshOpen('', '', '', '', '') + qsh + iXmlNodeQshClose();
-    }
-    return xml;
-  }
-
-  static generateShXML(sh, options) {
-    let xml;
-    if (options) {
-      xml = iXmlNodeShOpen(options.rows, options.hex, options.before, options.after, options.error)
-            + sh + iXmlNodeShClose();
-    } else {
-      xml = iXmlNodeShOpen('', '', '', '', '') + sh + iXmlNodeShClose();
-    }
-    return xml;
-  }
-}
-
-// DEPRECATED FUNCTIONS BELOW. WILL BE REMOVED SOMETIME AFTER v1.0
-
-/**
- * @description creates cmd XML
- * @param {string} cmd
- * @param {object} [options]
- * @returns {string} - the generated XML for the CL command
- */
-const iCmd = (cmd, options) => {
-  let xexec = 'cmd';
+const generateCmdXml = (cmd, options) => {
+  const xexec = (cmd.indexOf('?') > 0) ? 'rexx' : 'cmd'; // if need return value, use 'rexx'
   let xml;
-
-  if (cmd.indexOf('?') > 0) { // If need return value
-    xexec = 'rexx';
-  }
 
   if (options) {
     xml = iXmlNodeCmdOpen(options.exec, options.hex, options.before, options.after, options.error)
@@ -89,13 +39,7 @@ const iCmd = (cmd, options) => {
   return xml;
 };
 
-/**
- * @description creates qsh XML
- * @param {string} qsh
- * @param {object} [options]
- * @returns {string} - the generated XML for the qsh command
- */
-const iQsh = (qsh, options) => {
+const generateQshXml = (qsh, options) => {
   let xml;
   if (options) {
     xml = iXmlNodeQshOpen(options.rows, options.hex, options.before, options.after, options.error)
@@ -106,13 +50,7 @@ const iQsh = (qsh, options) => {
   return xml;
 };
 
-/**
- * @description creates sh XML
- * @param {string} sh
- * @param {object} [options]
- * @returns {string} - the generated XML for the sh command
- */
-const iSh = (sh, options) => {
+const generateShXml = (sh, options) => {
   let xml;
   if (options) {
     xml = iXmlNodeShOpen(options.rows, options.hex, options.before, options.after, options.error)
@@ -123,7 +61,6 @@ const iSh = (sh, options) => {
   return xml;
 };
 
-module.exports.CommandGenerator = CommandGenerator;
-module.exports.iCmd = iCmd;
-module.exports.iQsh = iQsh;
-module.exports.iSh = iSh;
+module.exports.generateCmdXml = generateCmdXml;
+module.exports.generateQshXml = generateQshXml;
+module.exports.generateShXml = generateShXml;
