+++ lib/itoolkit.js
@@ -20,46 +20,42 @@
  * main entry point for itoolkit package
  */
 
-const {
-  ProgramCall,
-  iPgm, // deprecated
-} = require('./ProgramCall');
-
-const {
-  CommandGenerator,
-  iCmd, // deprecated
-  iQsh, // deprecated
-  iSh, // deprecated
-} = require('./CommandGenerator');
-
-const {
-  SqlCall,
-  iSql, // deprecated
-} = require('./SqlCall');
-
-const {
-  Connection,
-  iConn, // deprecated
-} = require('./Connection');
-
+const { ProgramCall } = require('./ProgramCall');
+const { SqlCall } = require('./SqlCall');
+const { Connection } = require('./Connection');
+const { Toolkit } = require('./Toolkit');
 const { xmlToJson } = require('./utils');
+
 const {
-  Toolkit,
-  iDataQueue, // deprecated
-  iNetwork, // deprecated
-  iObj, // deprecated
-  iProd, // deprecated
-  iUserSpace, // deprecated
-  iWork, // deprecated
-} = require('./Toolkit');
+  generateCmdXml,
+  generateQshXml,
+  generateShXml,
+} = require('./commands');
+
+const {
+  iPgm,
+  iSql,
+  iConn,
+  iDataQueue,
+  iNetwork,
+  iObj,
+  iProd,
+  iUserSpace,
+  iWork,
+  iCmd,
+  iQsh,
+  iSh,
+} = require('./Deprecated');
 
 module.exports = {
-  CommandGenerator,
   ProgramCall,
   SqlCall,
   Connection,
   Toolkit,
   xmlToJson,
+  generateCmdXml,
+  generateQshXml,
+  generateShXml,
   // deprecated exports below, replaced by functionality above
   iCmd,
   iConn,
@@ -74,15 +70,3 @@ module.exports = {
   iUserSpace,
   iWork,
 };
-
-
-// Eventually, file will look something like this:
-
-// const { ProgramCall } = require('./ProgramCall');
-// const { CommandGenerator } = require('./CommandGenerator);
-// const { Connection } = require('./Connection);
-// const { Toolkit } = require('./Toolkit');
-// const { xmlToJson } = require('./utils');
-//
-// module.exports = { ... }
-//
